requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.5
psycopg2-binary==2.9.1
alembic==1.7.1
tensorflow==2.9.1
scikit-learn==0.24.2
xgboost==1.5.0
pandas==1.3.3
numpy==1.21.2
python-dotenv==0.19.0
requests==2.26.0
aiohttp==3.7.4
pytest==6.2.5
black==21.7b0
flake8==3.9.2
mypy==0.910
python-jose[cryptography]==3.3.0
boto3==1.18.50
stripe==2.60.0
sendgrid==6.8.1
redis==3.5.3
celery==5.1.2 